feat: hide the password on `ipmitool` calls
Related to talos-systems/sidero#483

This change uses a simple trick to make `ipmitool` read the actual
password from the removed temporary file only accessible via a file
descriptor.

This makes sure actual password will never be part of the command line
(visible in `ps`) and in the error logs of Sidero.

Signed-off-by: Andrey Smirnov <andrey.smirnov@talos-systems.com>